feat(api): user should be able to update the price of user's orders - modify test files into various specific files - add test for making purchase order - add route for purchase order - add controller method for purchase order - add model method for purchase order
[Delivers 166103227]
