+++ src/server.js
@@ -12,6 +12,12 @@ app.use(express.json());
 app.use(express.urlencoded({ extended: false }));
 app.use('/api/v1', router1);
 
+
+function redirectUnmatched(req, res) {
+  return res.status(400).json({ status: 400, error: 'Malformed Path' });
+}
+app.use(redirectUnmatched);
+
 const server = app.listen(portNumber);
 
 console.log('app running on port ', portNumber);
